Refactor: getButtonText 폴더 lib로 이동
